Made all the required changes reccomended by LFA
