+++ Server/.travis.yml
@@ -1,5 +1,2 @@
-language: node_js
-node_js:
-  - "node"
 script:
   - npm test
