+++ Server/server/models/index.js
@@ -1,12 +1,12 @@
 
 
-const fs = require('fs');
-const path = require('path');
-const Sequelize = require('sequelize');
+import fs from 'fs';
+import path from 'path';
+import Sequelize from 'sequelize';
 
 const basename = path.basename(__filename);
 const env = process.env.NODE_ENV || 'development';
-const config = require('${__dirname}/../config/config.json')[env];
+const config = require(`${__dirname}/../config/config.json`)[env];
 
 const db = {};
 
@@ -35,4 +35,4 @@ Object.keys(db).forEach(modelName => {
 db.sequelize = sequelize;
 db.Sequelize = Sequelize;
 
-module.exports = db;
+export default db;
