+++ Server/tests/test.js
@@ -1,6 +1,6 @@
 
 import { describe, it } from 'mocha';
-import chai, ( expect ) from 'chai';
+import chai, { expect } from 'chai';
 import chaiHttp from 'chai-http';
 import app from '../src/app';
